change navlink design
