+++ package.json
@@ -13,6 +13,7 @@
     "@reduxjs/toolkit": "^1.9.5",
     "antd": "^5.10.3",
     "axios": "^1.4.0",
+    "clsx": "^2.1.0",
     "cookiejs": "^2.1.2",
     "framer-motion": "^10.12.18",
     "rc-progress": "^3.4.2",
@@ -20,7 +21,8 @@
     "react-dom": "^18.2.0",
     "react-icons": "^4.10.1",
     "react-redux": "^8.1.1",
-    "react-router-dom": "^6.14.0"
+    "react-router-dom": "^6.14.0",
+    "tailwind-merge": "^2.2.2"
   },
   "devDependencies": {
     "@types/react": "^18.0.37",
@@ -33,6 +35,7 @@
     "eslint-plugin-react-refresh": "^0.3.4",
     "postcss": "^8.4.24",
     "tailwindcss": "^3.3.2",
-    "vite": "^4.3.9"
+    "vite": "^4.3.9",
+    "vite-jsconfig-paths": "^2.0.1"
   }
 }
