+++ src/components/links/link.css
@@ -1,29 +1,30 @@
 .nav-link.active {
+  color: white;
   border-bottom: 1px solid;
-  color: white
 }
 
+/*
 .nav-link:before{
-  content: "";
   position: absolute;
   bottom: 0;
   left: 0;
+  z-index: -1;
   width: 100%;
   height: 0;
   background: linear-gradient(to top, rgba(255, 255, 255, 0.3) ,  rgba(255, 255, 255, 0.15) , transparent);
-  z-index: -1;
+  content: "";
   transition: 0.3s;
 }
 
 .nav-link.active:before, .nav-link:hover:before{
   height: 100%;
-}
+}*/
 
-@media screen and (max-width: 1024px) {
+/*@media screen and (max-width: 1024px) {
   .nav-link.active {
     border-bottom: none;
     border-left: 4px solid;
     color: black;
     background: rgb(0, 0, 0, 0.3);
   }
-}
+}*/
